Bug fix for issue #25
What: Source ICFG: Missing arbitrary edges b/w successive lines in a method

Why:
this was happening because edges were being created between nodes that
were not part of the graph. This would happen in case the node with a
certain label already existed and for we tried to create the node
with the same label, again; and since a node with the label already
existed the new node would be ignored by the graph. But, the edge was
getting created with the new node, not the original node.

Fix:
This fix ensures that if a node with a given lable already exsits,
then a new node is not created, and the existing/orignal node is
used instead, esp. for creating new edges.