Blinky Statik: Source ICFG: Avoid self loops
What: This change avoids any self loops in the source ICFG, where
self loops are edges with the same source and target.

Why:
- This can happen if a source code statement (in Java) is
translated to two or more successive Jimple (or other IR) statements.
Thereby, resulting in the edges between the Jimple statements, which
belong to the same source statement, to get translated to an edge in
the soruce ICFG where the source and traget nodes are the same.
- Also having such statements is not particularly useful in the
source level ICFG

Fix: simply not adding an edge between two node if they are the same.