Blinky Statik: Call Graph Management with Dedicated API (Ref. Issue #29)
What:
This change introduces a clear API to manage the static call
graphs that is created with the help of Soot. This is supported
with a new CallGraphManager class that supports the following
methods in its API:
- getCallgraph
- getEntryPoints
- addEntryPoint
- methodExists
- getMethodCallers
- getMethodCallFrom
- getStrongComponents
- getBottomupTopology
- getDfsTraversal
- getBfsTraversal
- printGraph
The new CallGraphManager is housed in a new package called
analysis/statik/calls. This new package also houses the
StatikCallGraphBuilder (earlier in analysis/statik) that
talks to Soot directly to create the call graph.

Why:
This provides a clear API surface to deal with creation and
management of static call graphs, which are turning out to be
a quintessential data-structure for any form of static analysis.

Testing: mostly manual testing based on the CallGraphSubject in
the test source folder. This commit, did not change the call
graph and ICFG that was produced for the CallGraphSubject.