Blinky Statik: Analysis Config: Call Graph Algorithm option
What: This provides a new config option to specify the algorithm
to compute the call-graph. Specifically, we now have 2 algorithms
that can be used - CHA (call heirarchy analysis) and SPARK (which
is a form of CFA-based analysis that accounts for calling context
-- i think).

Why: This allows the Statik sub-module's users to generate different
call-graphs with trade-offs between speed and accuracy (CHA is faster,
SPARK is more accurate).

Testing: Tried out the call-graph option with `cha` and `spark` as
values for that option and manually inspected the resulting call-
graph.