[Blinky Config] Introducing Blinky Config Refer Issue #38
What: There has been a long-standing need to provide a way of
specifying verbose configurations for a client java-agent,
like Blinky Core, without changing the client's source code.

Why: the current mode of specifying configuration values esp.
for inclusion/exclusion lists, is to do so by storing them in
Java source files (currently Deputy.java) that are part of the
compiled build, and changing the configurable properties in the
source files directly.

Solution:
- This sub-module enables the specification and modification of
configuration values in a Yaml file, and *not* a Java source file.
- The sub-module accepts three inputs: (a) the Yaml file with the
configuration values; (b) the classfile to which the configurable
values is compiled to; (c) the location of the compiled output where
classfile is stored.
- This sub-module will compile the values in the Yaml config-file to
generate/modify appropriate bytecode in the input Java bytecode 
classfile that can be used by the client java-agent's code to access
user-provided configurations.

Testing: See unit tests; also performed manual tests to confirm that
the bytecode compilation is being performed as expected.