Executable Statik Analysis Pipelines (Ref. Issue #39)
What: With this change - this commit and the commit before this,
we are now able to weave multiple executable blocks that
implement some form of static analysis, in a single pipeline.

Why: This basically sets up different analyses that devs may want
to implement and execute on the same code/system under analysis.
Essentially, no more changing the main method in Statik class to
change the kind of static analysis to be run on the code under
analysis ... All changes to the config file.

Details:

- AnalysisBlock Class
Specifically, all analysis code will be implemented in classes that
extend the `AnalysisBlock` class in the `org.spideruci.analysis.statik.blocks`
package. This class has an `execute` method that is required by
the sub-classes to implement. This method takes as input an instance
of the `ClassGraphManager`. This gives all analysis-block access
to the call graph of code/system under analysis. All these classes
are implemented under the package `org.spideruci.analysis.statik.blocks`.

- Specifying AnalysisBlocks in Config File
The AnalysisBlocks are executed when they are specified in the con-
fig file. If the blocks are to be considered for execution, they
should be prepended with the dollar sign (`$`). Further, they must
be specified with the fully qualified names of the sub-class (i.e.,
with the package name) that is implementing the `AnalysisBlock`
class.

- The Executor Look in Statik.java
An exec loop in Statik will create an instance for each sub-class
specified in the Config file. The execute method on each of those
instances will then be executed, while passing along the Call Graph
manager of the code under analysis to each `execute` invocations.